+++ InventoryManagementSystem/InventoryManagementSystem/dataAccess/GraphicCardDataAccess.cs
@@ -100,7 +100,7 @@ namespace InventoryManagementSystem.dataAccess
             graphicCard.Description = reader.GetValue(1).ToString();
             graphicCard.ClockRate = Double.Parse(reader.GetValue(2).ToString());
             graphicCard.Model = reader.GetValue(3).ToString();
-            graphicCard.Memory = Int32.Parse(reader.GetValue(4).ToString());
+            graphicCard.Memory = ulong.Parse(reader.GetValue(4).ToString());
             graphicCard.Producer = producerDataAccess.GetEntityById<Producer>(Int32.Parse(reader.GetValue(5).ToString()));
             graphicCard.PhysicalInterfaces = this.GetPhysicalInterfaces(graphicCard);
 
@@ -126,7 +126,7 @@ namespace InventoryManagementSystem.dataAccess
             while (reader.Read())
             {
                 PhysicalInterface physicalInterface = physicalInterfaceDataAccess.GetEntityById<PhysicalInterface>(Int32.Parse(reader.GetValue(1).ToString()));
-                int count = Int32.Parse(reader.GetValue(2).ToString());
+                uint count = uint.Parse(reader.GetValue(2).ToString());
                 physicalInterfaces.Add(new PhysicalInterfaceWithCount(physicalInterface, count));
             }
